Changed behavior
Turns out that deltaTime can greatly fluctuate if the tab was paused or in background for a long time.
If this happens while emission burst is set, you'll see a whole bunch of particles spawning which isn't good in terms of performance.
Therefore, curTime value will be reset using modulus instead of subtraction.